@@ -1,7 +1,7 @@
 package com.arhohuttunen
 
-import javax.validation.constraints.Min
-import javax.validation.constraints.NotBlank
+import jakarta.validation.constraints.Min
+import jakarta.validation.constraints.NotBlank
 
 data class Person(
     @field:Min(1)
